Fix layout bug.
The old CSS caused issues on mobile browsers when pages had code blocks or images wider than the viewport. The pre block or image would take up the page width upto the max, but the text width would contract.

I have not understood the reason for this behavior, but handling overflow on pre and setting image max width as a percentage fixes the issue.